test(664): relocate 10 real-DB/file tests to the integration tier
These 10 files genuinely exercise REAL-DB / on-disk behavior that the
unit-tier in-memory redirect erases — journal_mode/WAL pragmas, file
locking, file-existence on disk, migrations, external opencode/claude DB
readers, and the purity guard itself. They cannot run under the :memory:
pool, so they move from the unit tier (tests/) to the integration tier
(tests/integration/), which leaves the redirect off and opens real files.

Also updates the lint-tests-isolation allowlist entry for the relocated
fixtures-stash-load test path.
