+++ tests/integration/commands/improve/improve-db-locking.test.ts
@@ -24,16 +24,16 @@ import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 import fs from "node:fs";
 import path from "node:path";
 
-import { akmImprove, runImproveMaintenancePasses } from "../../../src/commands/improve/improve";
-import { loadConfig, saveConfig } from "../../../src/core/config/config";
-import { readEvents } from "../../../src/core/events";
-import { insertEvent, openStateDatabase } from "../../../src/core/state-db";
-import { probeIndexWriterLease } from "../../../src/indexer/index-writer-lock";
-import { akmIndex } from "../../../src/indexer/indexer";
-import type { MemoryInferenceResult } from "../../../src/indexer/passes/memory-inference";
-import type { StalenessDetectionResult } from "../../../src/indexer/passes/staleness-detect";
-import type { Database } from "../../../src/storage/database";
-import { type IsolatedAkmStorage, makeSandboxDir, withIsolatedAkmStorage } from "../../_helpers/sandbox";
+import { akmImprove, runImproveMaintenancePasses } from "../../../../src/commands/improve/improve";
+import { loadConfig, saveConfig } from "../../../../src/core/config/config";
+import { readEvents } from "../../../../src/core/events";
+import { insertEvent, openStateDatabase } from "../../../../src/core/state-db";
+import { probeIndexWriterLease } from "../../../../src/indexer/index-writer-lock";
+import { akmIndex } from "../../../../src/indexer/indexer";
+import type { MemoryInferenceResult } from "../../../../src/indexer/passes/memory-inference";
+import type { StalenessDetectionResult } from "../../../../src/indexer/passes/staleness-detect";
+import type { Database } from "../../../../src/storage/database";
+import { type IsolatedAkmStorage, makeSandboxDir, withIsolatedAkmStorage } from "../../../_helpers/sandbox";
 
 let storage: IsolatedAkmStorage;
 const extraCleanups: Array<() => void> = [];
