+++ tests/integration/session-logs-providers.test.ts
@@ -12,10 +12,10 @@ import { afterEach, describe, expect, test } from "bun:test";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
-import { ClaudeCodeProvider } from "../src/integrations/harnesses/claude/session-log";
-import { OpenCodeProvider } from "../src/integrations/harnesses/opencode/session-log";
-import { extractInlineRefMentions } from "../src/integrations/session-logs/inline-refs";
-import { openDatabase } from "../src/storage/database";
+import { ClaudeCodeProvider } from "../../src/integrations/harnesses/claude/session-log";
+import { OpenCodeProvider } from "../../src/integrations/harnesses/opencode/session-log";
+import { extractInlineRefMentions } from "../../src/integrations/session-logs/inline-refs";
+import { openDatabase } from "../../src/storage/database";
 
 const tempDirs: string[] = [];
 function makeTempDir(prefix: string): string {
